requirements.txt
transformers==4.56.2
datasets>=2.21.0
accelerate==0.34.2
peft==0.11.1
bitsandbytes==0.47.0
unsloth==2025.7.2
trl==0.9.6
pyyaml==6.0.3
